improve error message of throw-to-outline when no outline present
****** behavior

when tro-refile from a text buffer to a non-dired buffer:
the error occurs if the first heading in the visible region is not an outline heading.
returning message:
"outline-back-to-heading: Before first heading"
because (outline-hide-subtree) runs without first verifying that it is on an outline heading.

****** this is undesirable for two reasons:

1. the error message is vague
2. there may be a valid outline below the first visible line, leaving the user unsure how to correct the problem.

****** solution

include code to verify point is on a heading, or try to advance to find one if it is not, and if that fails, return an error.