replace needless instance of (other-window -1)
> You're using both (select-window (previous-window)) and (other-window -1) within the same function -- do these do the same thing? Worth settling on one?

They don't do exactly the same thing. other-window is an interactive function. I use it for the last window selection step, because it triggers golden-ratio-mode to resize the windows appropriately. This doesn't need to happen while the function is working.

That said, one of the two instances is unnecessary, since the starting window is the same as the ending window, so removed it.